+++ dog-home/src/main/java/com/dianping/dog/alarm/rule/RuleManager.java
@@ -8,7 +8,7 @@ public interface RuleManager {
 
 	List<Rule> getRules();
 	
-	Rule getRuleById(long id);
+	Rule getRuleById(int id);
 	
 	boolean addRule(RuleEntity entity);
